Move 'Manager' to seperate include
