+++ code/Color.h
@@ -31,97 +31,8 @@ struct hsv {
     float v;       // percent
 };
 
-hsv rgb2hsv(const rgb& in) {
-    hsv out;
-    float min, max, delta;
+hsv rgb2hsv(const rgb& in);
 
-    min = in.r < in.g ? in.r : in.g;
-    min = min < in.b ? min : in.b;
-
-    max = in.r > in.g ? in.r : in.g;
-    max = max > in.b ? max : in.b;
-
-    out.v = max;                                // v
-    delta = max - min;
-    if (max > 0.0) {
-        out.s = (delta / max);                  // s
-    } else {
-        // r = g = b = 0                        // s = 0, v is undefined
-        out.s = 0.0;
-        out.h = std::numeric_limits<float>::quiet_NaN();    // its now undefined
-        return out;
-    }
-    if (in.r >= max)                    // > is bogus, just keeps compilor happy
-        out.h = (in.g - in.b) / delta;        // between yellow & magenta
-    else if (in.g >= max)
-        out.h = 2.0f + (in.b - in.r) / delta;  // between cyan & yellow
-    else
-        out.h = 4.0f + (in.r - in.g) / delta;  // between magenta & cyan
-
-    out.h *= 60.0f;                              // degrees
-
-    if (out.h < 0.0f)
-        out.h += 360.0f;
-
-    return out;
-}
-
-rgb hsv2rgb(const hsv& in) {
-    float hh, p, q, t, ff;
-    long i;
-    rgb out;
-
-    if (in.s <= 0.0) {       // < is bogus, just shuts up warnings
-        out.r = in.v;
-        out.g = in.v;
-        out.b = in.v;
-        return out;
-    }
-    hh = in.h;
-    if (hh >= 360.0)
-        hh = 0.0;
-    hh /= 60.0;
-    i = (long) hh;
-    ff = hh - i;
-    p = in.v * (1.0f - in.s);
-    q = in.v * (1.0f - (in.s * ff));
-    t = in.v * (1.0f - (in.s * (1.0f - ff)));
-
-    switch (i) {
-    case 0:
-        out.r = in.v;
-        out.g = t;
-        out.b = p;
-        break;
-    case 1:
-        out.r = q;
-        out.g = in.v;
-        out.b = p;
-        break;
-    case 2:
-        out.r = p;
-        out.g = in.v;
-        out.b = t;
-        break;
-
-    case 3:
-        out.r = p;
-        out.g = q;
-        out.b = in.v;
-        break;
-    case 4:
-        out.r = t;
-        out.g = p;
-        out.b = in.v;
-        break;
-    case 5:
-    default:
-        out.r = in.v;
-        out.g = p;
-        out.b = q;
-        break;
-    }
-    return out;
-}
+rgb hsv2rgb(const hsv& in);
 }
 }
